Add custom getrandom wrapper
We do not fall back to reading a potentially uninitialized /dev/urandom
anymore. Limiting `getrandom()` requests to 256 bytes each allows us to
report an uninitialized entropy pool to the user. This change also
removes a potentially fragile dependency on the standard library.

This is currently only implemented for Linux 3.17 and later. We might
add other operating systems later.